+++ include/D3D12Helper/D3D12Gpu/D3D12Gpu.hpp
@@ -4,7 +4,9 @@
 #include <D3D12Helper/D3D12Gpu/D3D12CpuImage.hpp>
 #include <D3D12Helper/D3D12Gpu/D3D12Mipmap.hpp>
 #include <D3D12Helper/D3D12Gpu/D3D12Resolve.hpp>
+#include <D3D12Helper/D3D12Gpu/D3D12State.hpp>
 #include <D3D12Helper/D3D12Gpu/D3D12TextureTransfer.hpp>
+#include <D3D12Helper/D3D12Gpu/D3D12View.hpp>
 #include <D3D12Helper/D3D12Framework/D3D12Resource.hpp>
 #include <D3D12Helper/D3D12Framework/D3D12DescriptorHandle.hpp>
 #include <D3D12Helper/D3D12Framework/D3D12DescriptorHeap.hpp>